refactor(api): rename getCombinedCourseByOrganizationId to getCombinedCoursesByOrganizationId and update references
